Closes #43

Some cleanup before working on next set of features. Until now a lot of
work has be done and a quick checkup and cleanup need to be done

+++ tests/e2e/test_run.test.cpp
@@ -21,11 +21,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-#include "../../include/crossedfingers/internals/utils.hpp"
 #include "test_tools.hpp"
 
-#include <algorithm>
 #include <crossedfingers/test.h>
+#include <crossedfingers/utilities.h>
 
 using namespace crossedfingers;
 
@@ -33,7 +32,7 @@ describe(test_run, []() {
     it("Should have the same count of test between list and run", []() {
         const auto list      = run_with_args("list");
         const auto test_list = split(list, "\n");
-        assertThat(test_list.size()).isGreaterThanOrEqualTo(1);
+        assertThat(test_list.size()).isGreaterThanOrEqualTo(static_cast<unsigned int>(1));
 
         const auto run = run_with_args("run");
         assertThat(run).hasSubString("TEST COUNT: " + std::to_string(test_list.size()));
